+++ pkg/tuple/onrset_test.go
@@ -11,15 +11,15 @@ import (
 func TestONRSet(t *testing.T) {
 	set := NewONRSet()
 	require.True(t, set.IsEmpty())
-	require.Equal(t, uint32(0), set.Length())
+	require.Equal(t, uint64(0), set.Length())
 
 	require.True(t, set.Add(ParseONR("resource:1#viewer")))
 	require.False(t, set.IsEmpty())
-	require.Equal(t, uint32(1), set.Length())
+	require.Equal(t, uint64(1), set.Length())
 
 	require.True(t, set.Add(ParseONR("resource:2#viewer")))
 	require.True(t, set.Add(ParseONR("resource:3#viewer")))
-	require.Equal(t, uint32(3), set.Length())
+	require.Equal(t, uint64(3), set.Length())
 
 	require.False(t, set.Add(ParseONR("resource:1#viewer")))
 	require.True(t, set.Add(ParseONR("resource:1#editor")))
@@ -40,7 +40,7 @@ func TestONRSetUpdate(t *testing.T) {
 		ParseONR("resource:2#viewer"),
 		ParseONR("resource:3#viewer"),
 	})
-	require.Equal(t, uint32(3), set.Length())
+	require.Equal(t, uint64(3), set.Length())
 
 	set.Update([]*core.ObjectAndRelation{
 		ParseONR("resource:1#viewer"),
@@ -49,7 +49,7 @@ func TestONRSetUpdate(t *testing.T) {
 		ParseONR("resource:1#admin"),
 		ParseONR("resource:1#reader"),
 	})
-	require.Equal(t, uint32(7), set.Length())
+	require.Equal(t, uint64(7), set.Length())
 }
 
 func TestONRSetIntersect(t *testing.T) {
@@ -70,8 +70,8 @@ func TestONRSetIntersect(t *testing.T) {
 		ParseONR("resource:1#reader"),
 	})
 
-	require.Equal(t, uint32(2), set1.Intersect(set2).Length())
-	require.Equal(t, uint32(2), set2.Intersect(set1).Length())
+	require.Equal(t, uint64(2), set1.Intersect(set2).Length())
+	require.Equal(t, uint64(2), set2.Intersect(set1).Length())
 }
 
 func TestONRSetSubtract(t *testing.T) {
@@ -92,8 +92,8 @@ func TestONRSetSubtract(t *testing.T) {
 		ParseONR("resource:1#reader"),
 	})
 
-	require.Equal(t, uint32(1), set1.Subtract(set2).Length())
-	require.Equal(t, uint32(4), set2.Subtract(set1).Length())
+	require.Equal(t, uint64(1), set1.Subtract(set2).Length())
+	require.Equal(t, uint64(4), set2.Subtract(set1).Length())
 }
 
 func TestONRSetUnion(t *testing.T) {
@@ -114,8 +114,8 @@ func TestONRSetUnion(t *testing.T) {
 		ParseONR("resource:1#reader"),
 	})
 
-	require.Equal(t, uint32(7), set1.Union(set2).Length())
-	require.Equal(t, uint32(7), set2.Union(set1).Length())
+	require.Equal(t, uint64(7), set1.Union(set2).Length())
+	require.Equal(t, uint64(7), set2.Union(set1).Length())
 }
 
 func TestONRSetWith(t *testing.T) {
@@ -127,8 +127,8 @@ func TestONRSetWith(t *testing.T) {
 	})
 
 	added := set1.With(ParseONR("resource:1#editor"))
-	require.Equal(t, uint32(3), set1.Length())
-	require.Equal(t, uint32(4), added.Length())
+	require.Equal(t, uint64(3), set1.Length())
+	require.Equal(t, uint64(4), added.Length())
 }
 
 func TestONRSetAsSlice(t *testing.T) {
